fix: move StyledBlockquote to component
