+++ src/app/components/article/article-text-builder.js
@@ -1,29 +1,6 @@
 import React from 'react';
-import styled from 'react-emotion';
 import StyledLink from '../styled-link/styled-link';
-import { color, fontFamily, text, grid } from '../../styles';
-
-const StyledBlockquote = styled('blockquote')`
-  display: flex;
-  color: ${color.kiev};
-  margin: 0 ${grid.gutter.xl} 0 0;
-  font-family: ${fontFamily.sans};
-  font-size: ${text.sizeStep['1']};
-  font-weight: bold;
-  line-height: ${text.lineHeight.basic};
-
-  &:before {
-    content: '“';
-    display: block;
-    position: relative;
-    top: 6px;
-    width: ${grid.gutter.l};
-    padding-right: calc(${grid.spacing.mouse} * 1.5);
-    color: ${color.economist};
-    font-family: ${fontFamily.serif};
-    font-size: ${text.sizeStep['7']};
-  }
-`;
+import StyledBlockquote from '../styled-block-quote/styled-block-quote';
 
 const buildArticleText = (childrenData = [], path = '') => {
   const getHTMLTag = (name, children, key) => {
@@ -54,7 +31,7 @@ const buildArticleText = (childrenData = [], path = '') => {
             return <br key={key} />;
           case 'blockquote':
             return (
-              <StyledBlockquote>
+              <StyledBlockquote key={key}>
                 {buildArticleText(children, key)}
               </StyledBlockquote>
             );
